+++ CHANGELOG.md
@@ -12,8 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `maxWidthPixels` prop to image element
 - `useFormSubmissionDuration` hook
 - `formStyling` prop to `OneBlinkFormBaseProps`
-- `useApplyUserDefinedFormStyling` hook
-- `objectEntriesAsKeyValuePairs` function
+- `useApplyUserDefinedStyling` hook
+- `objectEntriesAsTypedKeyValuePairs` function
 
 ## [8.13.2] - 2025-12-04